Implemented a static track to display four quarter notes after the start button is pressed.
There needs to be a dynamically updating display to represent the changing notes needed for different bars. I was thinking of using an equation that would look for the difference between the previous note and the upcoming note and use the enums to determine what image should be displayed based on the bpm.