Applied small fixes after Willey's meeting.
Fixed the spelling errors on the TrackSelect screen and in the footer. Relocated to footer to only the Home screen. Changed the Dr. Willey link to Help in the footer. Changed the text on the 'Start Timer' button to 'Start'. Changed the blurb on the Home screen. Also, attempted to fix the counting before the track played.